refactor(server): Routing, Serialization 파일을 plugins 폴더로 이동
Ktor의 권장 구조인 플러그인 기반 아키텍처의 용이한 유지보수를 위해
plugins 패키지로 분리하여 유지보수를 용이하게 했습니다.
