+++ server/src/main/java/com/example/server/Application.kt
@@ -1,5 +1,6 @@
 package com.example.server
 
+import com.example.server.plugins.configureRouting
 import io.ktor.server.application.*
