+++ server/src/main/java/com/example/server/plugins/Routing.kt
@@ -1,4 +1,4 @@
-package com.example.server
+package com.example.server.plugins
 
 import com.example.server.model.Component
 import com.example.server.model.ComponentRepository
