+++ server/src/main/java/com/example/server/plugins/Serialization.kt
@@ -1,11 +1,16 @@
-package com.example.server.plugins
+package com.example.server
 
-import io.ktor.serialization.kotlinx.json.*
-import io.ktor.server.application.*
-import io.ktor.server.plugins.contentnegotiation.*
+import com.example.server.model.ComponentRepository
+import io.ktor.serialization.kotlinx.json.json
+import io.ktor.server.application.Application
+import io.ktor.server.application.install
+import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.server.response.respond
+import io.ktor.server.routing.get
+import io.ktor.server.routing.routing
 
 fun Application.configureSerialization() {
     install(ContentNegotiation) {
-        json() // kotlinx.serialization을 JSON 엔진으로 사용
+        json()
     }
 }
